+++ internal/controller/proxy_controller.go
@@ -17,17 +17,6 @@ import (
 	"github.com/google/go-querystring/query"
 )
 
-type UnauthorizedQuery struct {
-	Username string `url:"username"`
-	Resource string `url:"resource"`
-	GroupErr bool   `url:"groupErr"`
-	IP       string `url:"ip"`
-}
-
-type RedirectQuery struct {
-	RedirectURI string `url:"redirect_uri"`
-}
-
 type AuthModuleType int
 
 const (
